New scripts and bugfix
Added python script that computes the difference between results, another script that convert the results in text format and a gnuplot script to plot them. Added gnuplot script to plot in linear scale only the first 20 fm of the system evolution. Fixed a bug in the definition of nucleons, now only protons and neutrons, not their antiparticles